tree: Reuse flex tree fields in most cases (#21570)
## Description

Currently flex-tree fields tend to be leaked until the nodes they are
part of are disposed (if ever) or the context is disposed.

Due to this, allocating tons of separate objects for the same field is
really bad. This change reuses flex tree fields in most cases,
mitigating this issue.

This was observed to make bubble bench >10x faster, however most of that
win was due to avoiding the O(number of leaked copies of the field)
creation cost in the event subscription logic which is being removed
independently.

(cherry picked from commit e8d9abebc62bd18e722fdc5f670f7b800259a40d)